Complete 64-bit type safety improvements after cast analysis
This commit addresses all remaining type safety issues identified in the
comprehensive (int) cast analysis (see INT_CAST_SAFETY_ANALYSIS.md).

Critical fixes (Priority 1):
- Change TFCstrcpy return type from int to ptrdiff_t (tfed.c:2263)
- Change TFCstrncpy return type from int to ptrdiff_t (tfed.c:2279)
- Update function declarations in tfed_i.h
- Remove unsafe (int) casts from pointer arithmetic

Medium priority fixes (Priority 2):
- Change 'length' variable from int to size_t in addCtrlChar() (tfed.c:2748)
- Update all string length calculations to use size_t (10 locations)
- Change line/column parameters from int to long in cl_txtDisp.c
  * meth_txtDisp_deHighLight (lines 659-660)
  * meth_txtDisp_highLight (lines 788-789)

Low priority improvements:
- Add safe strlen wrapper for X11 API calls (tfed_i.h:81-85)
- Use size_t for strlen result before casting to int (cl_PS.c:143-148)
- Change extractWord parameters from int to long (mystrings.h:178, mystrings.c:992)
- Remove unnecessary casts when calling extractWord (cl_generic.c:3581)

Impact:
- Prevents potential data loss with large text buffers (>2GB)
- Allows text editors to handle files with >2 billion lines
- Improves type safety for string operations
- All changes are non-breaking (binary compatible)

Testing:
- Full clean build successful
- No new compiler warnings introduced
- Binary size unchanged (1.4M for both vw and viola)

Related documentation: INT_CAST_REVIEW_README.md

+++ src/viola/cl_txtDisp.c
@@ -656,8 +656,8 @@ int argc;
 Packet argv[];
 {
     TFStruct* tf = updateEStrUser(self);
-    int ln = (int)PkInfo2Int(&argv[0]);
-    int cn = (int)PkInfo2Int(&argv[1]);
+    long ln = PkInfo2Int(&argv[0]);
+    long cn = PkInfo2Int(&argv[1]);
 
     lineFlagSet(tf, ln, cn, MASK_REVERSE, -1);
     if ((ln >= tf->screen_row_offset) && (ln < tf->screen_row_offset + tf->num_of_lines)) {
@@ -785,8 +785,8 @@ int argc;
 Packet argv[];
 {
     TFStruct* tf = updateEStrUser(self);
-    int ln = (int)PkInfo2Int(&argv[0]);
-    int cn = (int)PkInfo2Int(&argv[1]);
+    long ln = PkInfo2Int(&argv[0]);
+    long cn = PkInfo2Int(&argv[1]);
 
     lineFlagSet(tf, ln, cn, MASK_REVERSE, 1);
     if ((ln >= tf->screen_row_offset) && (ln < tf->screen_row_offset + tf->num_of_lines)) {
